refactor: use drop down
