+++ src/renderer/src/modules/entry-content/components/EntryReadHistory.tsx
@@ -1,6 +1,12 @@
 import { useWhoami } from "@renderer/atoms/user"
 import { Avatar, AvatarFallback, AvatarImage } from "@renderer/components/ui/avatar"
 import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@renderer/components/ui/dropdown-menu/dropdown-menu"
+import {
   Tooltip,
   TooltipContent,
   TooltipPortal,
@@ -13,8 +19,7 @@ import { useUserById } from "@renderer/store/user"
 import { LayoutGroup, m } from "framer-motion"
 import { memo, useEffect, useState } from "react"
 
-import { usePresentUserProfileModal } from "../../../profile/hooks"
-import { useEntryReadHistoryModal } from "./hooks"
+import { usePresentUserProfileModal } from "../../profile/hooks"
 
 export const EntryReadHistory: Component<{ entryId: string }> = ({ entryId }) => {
   const me = useWhoami()
@@ -34,8 +39,6 @@ export const EntryReadHistory: Component<{ entryId: string }> = ({ entryId }) =>
       clearTimeout(timer)
     }
   }, [entryId])
-
-  const presentEntryReadHistoryModal = useEntryReadHistoryModal({ entryId })
 
   if (!entryHistory) return null
   if (!me) return null
@@ -57,32 +60,66 @@ export const EntryReadHistory: Component<{ entryId: string }> = ({ entryId }) =>
         entryHistory.readCount > 10 &&
         entryHistory.userIds &&
         entryHistory.userIds.length >= 10 && (
-          <Tooltip>
-            <TooltipTrigger className="no-drag-region relative cursor-pointer" asChild>
-              <button
-                onClick={() => {
-                  presentEntryReadHistoryModal()
-                }}
-                type="button"
-                style={{
-                  right: "80px",
-                  zIndex: 11,
-                }}
-                className="relative flex size-7 items-center justify-center rounded-full border border-border bg-muted ring ring-background"
-              >
-                <span className="text-[10px] font-medium text-muted-foreground">
-                  +{Math.min(entryHistory.readCount - 10, 99)}
-                </span>
-              </button>
-            </TooltipTrigger>
-            <TooltipPortal>
-              <TooltipContent side="top">More</TooltipContent>
-            </TooltipPortal>
-          </Tooltip>
+          <DropdownMenu>
+            <Tooltip>
+              <TooltipTrigger className="no-drag-region relative cursor-pointer" asChild>
+                <DropdownMenuTrigger asChild>
+                  <button
+                    type="button"
+                    style={{
+                      right: "80px",
+                      zIndex: 11,
+                    }}
+                    className="relative flex size-7 items-center justify-center rounded-full border border-border bg-muted ring ring-background"
+                  >
+                    <span className="text-[10px] font-medium text-muted-foreground">
+                      +{Math.min(entryHistory.readCount - 10, 99)}
+                    </span>
+                  </button>
+                </DropdownMenuTrigger>
+              </TooltipTrigger>
+              <TooltipPortal>
+                <TooltipContent side="top">More</TooltipContent>
+              </TooltipPortal>
+            </Tooltip>
+            <DropdownMenuContent>
+              {entryHistory.userIds
+                .filter((id) => id !== me?.id)
+                .slice(10)
+                .map((userId) => (
+                  <DropdownMenuItem key={userId}>
+                    <EntryRow userId={userId} />
+                  </DropdownMenuItem>
+                ))}
+            </DropdownMenuContent>
+          </DropdownMenu>
         )}
     </div>
   )
 }
+
+const EntryRow: Component<{ userId: string }> = memo(({ userId }) => {
+  const user = useUserById(userId)
+  const presentUserProfile = usePresentUserProfileModal("dialog")
+  if (!user) return null
+
+  return (
+    <button
+      type="button"
+      className="flex items-center gap-4"
+      onClick={() => {
+        presentUserProfile(userId)
+      }}
+    >
+      <Avatar className="aspect-square size-10 overflow-hidden rounded-full border border-border">
+        <AvatarImage src={user?.image || undefined} />
+        <AvatarFallback>{user.name?.slice(0, 2)}</AvatarFallback>
+      </Avatar>
+
+      {user.name && <p className="text-sm">{user.name}</p>}
+    </button>
+  )
+})
 
 const EntryUser: Component<{
   userId: string
